pyproject.toml
[tool.poetry]
name = "infrastructure"
version = "0.1.0"
description = ""
authors = ["Jack Stockley <34308937+jnstockley@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest-playwright = "^0.3.3"
pylint = "^2.17.3"
pytest = "^7.3.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.autohooks.plugins.pylint]
arguments = ["--rcfile=.pylintrc", "-s", "n"]